move pages to src
